Move pathnames into utils.
