+++ client/app/bundles/Events/components/Groups.jsx
@@ -1,5 +1,5 @@
-import React, { Component } from 'react';
 import axios from 'axios';
+import React, { Component } from 'react';
 
 import Group from './Group';
 import SortableHeader from './SortableHeader';
