+++ client/app/bundles/Events/components/Pagination.jsx
@@ -1,5 +1,5 @@
-import React, { Component } from 'react';
 import _ from 'lodash';
+import React, { Component } from 'react';
 
 import { Link } from 'react-router-dom';
 import { addParamToQuery } from '../utils';
