+++ client/app/bundles/Events/utils/index.jsx
@@ -1,3 +1,4 @@
 export * from './DateTimeUtils';
 export * from './QueryUtils';
 export * from './Permissions';
+export * from './Pathnames';
